+++ TDTChocolate/TestingAdditions/TDTRandomFixtures.m
@@ -1,6 +1,4 @@
 #import "TDTRandomFixtures.h"
-#import "../FoundationAdditions/TDTAssert.h"
-#import "../FoundationAdditions/TDTRandom.h"
 
 @implementation NSString (TDTRandomFixtures)
 
@@ -42,8 +40,6 @@
       case 2:
         randomObject = [NSArray tdt_randomArrayOfLength:1];
         break;
-      default:
-        TDTAssertFailure();
     }
     [result addObject:randomObject];
   }
@@ -63,7 +59,7 @@
 @implementation NSDate (TDTRandomFixtures)
 
 + (instancetype)tdt_randomDate {
-  NSTimeInterval randomTimeInterval = TDTRandomDouble() * arc4random();
+  NSTimeInterval randomTimeInterval = arc4random();
   return [self dateWithTimeIntervalSince1970:randomTimeInterval];
 }